chore: torch dataset refactored into its own C++ files
